refact: refatoração de relations

+++ src/modules/line-mark/line-mark.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateLineMarkDto } from './dto/create-line-mark.dto';
 import { UpdateLineMarkDto } from './dto/update-line-mark.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -29,26 +29,51 @@ export class LineMarkService {
   }
 
   async findOneLineMark(id: string): Promise<LineMarkEntity> {
-    return await this.lineMarkRepository.findOneBy({id: id})
+    try{
+      const lineMark = await this.lineMarkRepository.findOneBy({id: id});
+      if(lineMark === null){
+        throw new NotFoundException(`A marca com id: ${id} não foi encontrada.`)
+      }
+      return lineMark
+    }
+    catch(err){
+      throw new NotFoundException(`A marca com id: ${id} não foi encontrada.`)
+    }
   }
 
   async updateLineMark(id: string, updateLineMarkDto: UpdateLineMarkDto): Promise<IUpdatedData> {
-    await this.lineMarkRepository.update(id, updateLineMarkDto)
-    const updatedMark = await this.lineMarkRepository.findOneBy({id: id})
-    return{
-      updateData: updateLineMarkDto,
-      data: updatedMark,
-      message: 'Marca atualizada com sucesso.'
+    try{
+      const lineMark = await this.lineMarkRepository.findOneBy({id: id});
+      if(lineMark === null){
+        throw new NotFoundException(`A marca com id: ${id} não foi encontrada.`)
+      }
+      await this.lineMarkRepository.update(id, updateLineMarkDto)
+      const updatedMark = await this.lineMarkRepository.findOneBy({id: id})
+      return{
+        updateData: updateLineMarkDto,
+        data: updatedMark,
+        message: 'Marca atualizada com sucesso.'
+      }
+    }
+    catch(err){
+      throw new NotFoundException(`A marca com id: ${id} não foi encontrada.`)
     }
   }
 
   async deleteLineMark(id: string): Promise<IDeletedData> {
-    const selectedMark = await this.lineMarkRepository.findOneBy({id: id}) 
-    await this.lineMarkRepository.delete(id)
-
-    return {
-      data: selectedMark,
-      message: 'Marca deletada com sucesso.'
+    try{
+      const selectedMark = await this.lineMarkRepository.findOneBy({id: id}) 
+      if(selectedMark === null){
+        throw new NotFoundException(`A marca com id: ${id} não foi encontrada.`)
+      }
+      await this.lineMarkRepository.delete(id)  
+      return {
+        data: selectedMark,
+        message: 'Marca deletada com sucesso.'
+      }
+    }
+    catch(err){
+      throw new NotFoundException(`A marca com id: ${id} não foi encontrada.`)
     }
   }
 }
